feat: apply logic for Get Funds after divorce, MyContracts/Wedding
