+++ src/components/GetFundsModal/GetFundsModal.helpers.tsx
@@ -1,11 +1,11 @@
-export interface ITokenAddressField {
+export interface IGetFundsModalTokenAddressField {
   id: number;
   address: string;
 }
 
-export const fieldsHelper: ITokenAddressField[] = [
+export const fieldsHelper: IGetFundsModalTokenAddressField[] = [
   {
-    address: '0x324345',
+    address: '',
     id: 0,
   },
 ];
