+++ src/config/abi/index.ts
@@ -1,10 +1,10 @@
 export { default as bep20Abi } from './bep20';
 export { default as saleAbi } from './sale';
+
 export { default as tokenMintableFreezableAbi } from './tokenMintableFreezable';
 export { default as tokenMintableNonFreezableAbi } from './tokenMintableNonFreezable';
 export { default as tokenNonMintableFreezableAbi } from './tokenNonMintableFreezable';
 export { default as tokenNonMintableNonFreezableAbi } from './tokenNonMintableNonFreezable';
-export { default as lostKeyFactoryAbi } from './lostKeyFactory';
 
 export {
   default as crowdsaleNonSoftCappableBonusableAbi,
@@ -18,6 +18,9 @@ export {
 export {
   default as crowdsaleSoftCappableNonBonusableAbi,
 } from './crowdsaleSoftCappableNonBonusable';
-export { default as weddingFactoryAbi } from './weddingFactory';
 
+export { default as weddingFactoryAbi } from './weddingFactory';
+export { default as weddingAbi } from './wedding';
+
+export { default as lostKeyFactoryAbi } from './lostKeyFactory';
 export { default as lostKeyAbi } from './lostKey';
