+++ src/config/abi/wedding.ts
@@ -1,0 +1,210 @@
+import { AbiItem } from 'web3-utils';
+
+export default [
+  {
+    type: 'constructor',
+    stateMutability: 'nonpayable',
+    inputs: [
+      { type: 'address', name: '_firstPartner', internalType: 'address' },
+      { type: 'address', name: '_secondPartner', internalType: 'address' },
+      {
+        type: 'uint256',
+        name: '_decisionTimeWithdrawal',
+        internalType: 'uint256',
+      },
+      {
+        type: 'uint256',
+        name: '_decisionTimeDivorce',
+        internalType: 'uint256',
+      },
+      {
+        type: 'uint8',
+        name: '_percentageToProposingWhenDisputed',
+        internalType: 'uint8',
+      },
+    ],
+  },
+  {
+    type: 'event',
+    name: 'DivorceProposed',
+    inputs: [
+      {
+        type: 'address',
+        name: 'proposedBy',
+        internalType: 'address',
+        indexed: false,
+      },
+    ],
+    anonymous: false,
+  },
+  {
+    type: 'event',
+    name: 'WithdrawalProposed',
+    inputs: [
+      {
+        type: 'address',
+        name: 'token',
+        internalType: 'address',
+        indexed: false,
+      },
+      {
+        type: 'address',
+        name: 'receiver',
+        internalType: 'address',
+        indexed: false,
+      },
+      {
+        type: 'uint256',
+        name: 'amount',
+        internalType: 'uint256',
+        indexed: false,
+      },
+      {
+        type: 'address',
+        name: 'proposedBy',
+        internalType: 'address',
+        indexed: false,
+      },
+    ],
+    anonymous: false,
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [
+      { type: 'address', name: 'token', internalType: 'contract IERC20' },
+      { type: 'address', name: 'receiver', internalType: 'address' },
+      { type: 'uint256', name: 'amount', internalType: 'uint256' },
+      { type: 'address', name: 'proposedBy', internalType: 'address' },
+      { type: 'uint256', name: 'timestamp', internalType: 'uint256' },
+    ],
+    name: 'activeWithdrawalProposal',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'agreeWithDivorce',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'uint256', name: '', internalType: 'uint256' }],
+    name: 'decisionTimeDivorce',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'uint256', name: '', internalType: 'uint256' }],
+    name: 'decisionTimeWithdrawal',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'disputeDivorce',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'bool', name: '', internalType: 'bool' }],
+    name: 'divorceDisputed',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'address', name: '', internalType: 'address' }],
+    name: 'divorceProposedBy',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'uint256', name: '', internalType: 'uint256' }],
+    name: 'divorceTimestamp',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'executeWithdrawalProposal',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'address', name: '', internalType: 'address' }],
+    name: 'firstPartner',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'getFundsAfterDivorce',
+    inputs: [{ type: 'address[]', name: '_tokens', internalType: 'address[]' }],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'bool', name: '', internalType: 'bool' }],
+    name: 'partsTakenAfterDivorce',
+    inputs: [
+      { type: 'address', name: '', internalType: 'contract IERC20' },
+      { type: 'uint256', name: '', internalType: 'uint256' },
+    ],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'uint8', name: '', internalType: 'uint8' }],
+    name: 'percentageToProposingWhenDisputed',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'proposeDivorce',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'proposeWithdrawal',
+    inputs: [
+      { type: 'address', name: '_token', internalType: 'address' },
+      { type: 'address', name: '_receiver', internalType: 'address' },
+      { type: 'uint256', name: '_amount', internalType: 'uint256' },
+    ],
+  },
+  {
+    type: 'function',
+    stateMutability: 'nonpayable',
+    outputs: [],
+    name: 'rejectWithdrawalProposal',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'address', name: '', internalType: 'address' }],
+    name: 'secondPartner',
+    inputs: [],
+  },
+  {
+    type: 'function',
+    stateMutability: 'view',
+    outputs: [{ type: 'bool', name: '', internalType: 'bool' }],
+    name: 'withdrawalProposalPending',
+    inputs: [],
+  },
+] as AbiItem[];
